Make Reporter runnable in desktop
